creating abstract client class and interface; implementing invdvidual…

+++ Accounts/AbstractAccount.cs
@@ -4,11 +4,11 @@ namespace Bank.Accounts
 {
 	public abstract class AbstractAccount()
 	{
-		protected int _accountNumber;
+		protected long _accountNumber;
 		protected double _saldo;
 		protected List<Transaction> _transactions;
 		public double AccountMaintenance { get; set; }
-		public abstract int GetAccountNumber();
+		public abstract long GetAccountNumber();
 		public abstract double GetSaldo();
 		public abstract List<Transaction> GetTransactions();
 		public abstract string BankStatment();
